+++ addon/components/flexberry-simpleolv.js
@@ -1052,9 +1052,10 @@ export default folv.extend(
     },
 
     copyJSONContent(event) {
-      Ember.$('#OLVToolbarInfoContent').select();
+      let infoModalDialog = this.get('_infoModalDialog');
+      infoModalDialog.find('.olv-toolbar-info-modal-dialog-content textarea').select();
       let copied = document.execCommand('copy');
-      let oLVToolbarInfoCopyButton = Ember.$('#OLVToolbarInfoCopyButton');
+      let oLVToolbarInfoCopyButton = infoModalDialog.find('.olv-toolbar-info-modal-dialog-copy-button');
       oLVToolbarInfoCopyButton.get(0).innerHTML = this.get('i18n').t(copied ? 'components.olv-toolbar.copied' : 'components.olv-toolbar.ctrlc');
       oLVToolbarInfoCopyButton.addClass('disabled');
     },
@@ -2621,7 +2622,8 @@ export default folv.extend(
         buttonName: '...', // Button displayed name.
         buttonAction: '...', // Action that is called from controller on this button click (it has to be registered at component).
         buttonClasses: '...', // Css classes for button.
-        buttonTitle: '...' // Button title.
+        buttonTitle: '...', // Button title.
+        disabled: true, // The state of the button is disabled if `true` or enabled if `false`.
       }
       ```
 
@@ -2851,7 +2853,7 @@ export default folv.extend(
       infoModalDialog.modal('show');
     }
 
-    let oLVToolbarInfoCopyButton = Ember.$('#OLVToolbarInfoCopyButton');
+    let oLVToolbarInfoCopyButton = infoModalDialog.find('.olv-toolbar-info-modal-dialog-copy-button');
     oLVToolbarInfoCopyButton.get(0).innerHTML = this.get('i18n').t('components.olv-toolbar.copy');
     oLVToolbarInfoCopyButton.removeClass('disabled');
     return infoContent;